reorganize profile card files to reduce nesting
