+++ frontend/src/components/pages/profile/ProfileCard/variants/EventCard.tsx
@@ -1,12 +1,12 @@
 import { Grid, Typography } from "@material-ui/core";
 import Event from "@public/illustrations/Event.svg";
-import LinkAction from "./actions/LinkAction";
-import ProfileCardBase from "./ProfileCardBase";
+import LinkAction from "../LinkAction";
+import BaseCard from "../BaseCard";
 
 /** Displays a card on the profile page that links to events to which the user has signed up. */
 const EventCard: React.VFC = (props) => {
   return (
-    <ProfileCardBase
+    <BaseCard
       title="Arrangementer"
       Action={<LinkAction text="Se arrangementer" link="/events" />}
       image={Event}
@@ -20,7 +20,7 @@ const EventCard: React.VFC = (props) => {
           </Typography>
         </Grid>
       </Grid>
-    </ProfileCardBase>
+    </BaseCard>
   );
 };
