+++ frontend/src/components/pages/profile/ProfileCard/variants/FormCard.tsx
@@ -1,12 +1,12 @@
 import { Grid, Typography } from "@material-ui/core";
 import Form from "@public/illustrations/Form.svg";
-import LinkAction from "./actions/LinkAction";
-import ProfileCardBase from "./ProfileCardBase";
+import LinkAction from "../LinkAction";
+import BaseCard from "../BaseCard";
 
 /** Displays a card on the profile page that links to the user's submitted listing applications. */
 const FormCard: React.VFC = (props) => {
   return (
-    <ProfileCardBase
+    <BaseCard
       title="Vervsøknader"
       Action={<LinkAction text="Se verv" link="/listings" />}
       image={Form}
@@ -18,7 +18,7 @@ const FormCard: React.VFC = (props) => {
           <Typography variant="body2">Her kan du se en oversikt over alle dine innsendte søknader på verv.</Typography>
         </Grid>
       </Grid>
-    </ProfileCardBase>
+    </BaseCard>
   );
 };
