+++ frontend/src/components/pages/profile/ProfileCard/variants/OrdersCard.tsx
@@ -1,12 +1,12 @@
 import { Grid, Typography } from "@material-ui/core";
 import CreditCard from "@public/illustrations/CreditCard.svg";
-import LinkAction from "./actions/LinkAction";
-import ProfileCardBase from "./ProfileCardBase";
+import LinkAction from "../LinkAction";
+import BaseCard from "../BaseCard";
 
 /** Displays a card on the profile page that links to the user's orders. */
 const OrdersCard: React.VFC = (props) => {
   return (
-    <ProfileCardBase
+    <BaseCard
       title="Ordrehistorikk"
       Action={<LinkAction text="Se ordrehistorikk" link="/ecommerce" {...props} />}
       image={CreditCard}
@@ -20,7 +20,7 @@ const OrdersCard: React.VFC = (props) => {
           </Typography>
         </Grid>
       </Grid>
-    </ProfileCardBase>
+    </BaseCard>
   );
 };
